+++ database/migrations/2019_02_18_005125_create_user_detail_table.php
@@ -14,8 +14,8 @@ class CreateUserDetailTable extends Migration
     public function up()
     {
         Schema::create('user_detail', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
 
             $table->string('profile')->nullable();
